+++ index.ts
@@ -891,6 +891,16 @@ namespace Enumerable {
          */
         reverse(): IOrderedEnumerable<T>;
         /**
+         * Handles current items as numbers and calculates the n-th root for each item.
+         * 
+         * @param {number} [power] The power. Default: 2
+         * @param {boolean} [handleAsInt] Handle as integer values (true) or floats (false).
+         *                                Default: (false)
+         * 
+         * @return {IEnumerable<number>} The new sequence.
+         */
+        root(power?: number, handleAsInt?: boolean): IEnumerable<number>;
+        /**
          * Handles current items as float values and return their nearest values.
          * 
          * @return {IEnumerable<number>} The new sequence.
@@ -2357,6 +2367,19 @@ namespace Enumerable {
             
             return this.orderByDescending(() => {
                 return i++;
+            });
+        }
+        /** @inheritdoc */
+        public root(power?: number, handleAsInt?: boolean): IEnumerable<number> {
+            power = parseFloat( toStringSafe(power).trim() );
+            if (isNaN(power)) {
+                power = 2;
+            }
+
+            return this.select(x => {
+                return invokeForValidNumber(x,
+                                            y => Math.pow(y, 1 / power),
+                                            handleAsInt);
             });
         }
         /** @inheritdoc */